+++ integrations/terraform-mwi/provider/kubernetes_data_source_test.go
@@ -17,40 +17,78 @@ limitations under the License.
 package provider
 
 import (
+	"fmt"
+	"slices"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/knownvalue"
-	"github.com/hashicorp/terraform-plugin-testing/statecheck"
-	"github.com/hashicorp/terraform-plugin-testing/tfjsonpath"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+	corev1 "k8s.io/api/core/v1"
+
+	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/tool/teleport/testenv"
 )
 
 func TestAccKubernetesDataSource(t *testing.T) {
-	const config = `
+	log := utils.NewSlogLoggerForTests()
+	ctx := t.Context()
+
+	process, kubeMock := setupKubernetesHarness(t, log)
+	rootClient := testenv.MakeDefaultAuthClient(t, process)
+	_, pt := setupKubernetesAccessBot(ctx, t, rootClient)
+
+	config := fmt.Sprintf(`
 provider "teleportmwi" {
-  proxy_server = "example.com:3080"
-  join_method  = "gitlab"
-  join_token   = "example-token"
+  proxy_server = "%s"
+  join_method  = "kubernetes"
+  join_token   = "%s"
+  insecure     = true
 }
 
 data "teleportmwi_kubernetes" "example" {
   selector = {
-    name = "barry"
+    name = "%s"
   } 
 }
-`
+
+provider "kubernetes" {
+  host                   = data.teleportmwi_kubernetes.example.output.host
+  tls_server_name        = data.teleportmwi_kubernetes.example.output.tls_server_name
+  client_certificate     = data.teleportmwi_kubernetes.example.output.client_certificate
+  client_key             = data.teleportmwi_kubernetes.example.output.client_key
+  cluster_ca_certificate = data.teleportmwi_kubernetes.example.output.cluster_ca_certificate
+}
+
+resource "kubernetes_namespace" "ns" {
+  metadata {
+    name = "tf-mwi-ds-test"
+  }
+}
+`, process.Config.Proxy.WebAddr.String(), pt.GetName(), kubeClusterName)
+
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		ExternalProviders: map[string]resource.ExternalProvider{
+			"kubernetes": {
+				Source: "hashicorp/kubernetes",
+			},
+		},
 		Steps: []resource.TestStep{
 			{
 				Config: config,
-				ConfigStateChecks: []statecheck.StateCheck{
-					statecheck.ExpectKnownValue(
-						"data.teleportmwi_kubernetes.example",
-						tfjsonpath.New("output").AtMapKey("host"),
-						knownvalue.StringExact("Hello, barry!"),
-					),
+				Check: func(state *terraform.State) error {
+					// Check that the namespace was actually created!
+					containsCreatedNamespace := slices.ContainsFunc(
+						kubeMock.ListNamespaces().Items,
+						func(ns corev1.Namespace) bool {
+							return ns.Name == "tf-mwi-ds-test"
+						},
+					)
+					if !containsCreatedNamespace {
+						return fmt.Errorf("expected to find created namespace in TF state")
+					}
+					return nil
 				},
 			},
 		},
